Move out the reset code to a common module
so that it can be reused by multiple mode inference modules
